refactor(llm): Remove unused code
