+++ apps/ledger-live-desktop/src/renderer/modals/UpdateFirmwareModal/helpers/createFirmwareUpdateSteps.ts
@@ -1,4 +1,7 @@
-import { isDeviceLocalizationSupported, FirmwareUpdateContext } from "@ledgerhq/live-device-core";
+import {
+  isDeviceLocalizationSupported,
+  FirmwareUpdateContext,
+} from "@ledgerhq/live-common/device/index";
 import { DeviceModelId } from "@ledgerhq/devices";
 import StepResetDevice, { StepResetFooter } from "../steps/00-step-reset-device";
 import StepPrepare from "../steps/01-step-prepare";
