+++ apps/ledger-live-mobile/package.json
@@ -86,7 +86,6 @@
     "@ledgerhq/live-env": "workspace:^",
     "@ledgerhq/live-nft": "workspace:^",
     "@ledgerhq/live-nft-react": "workspace:^",
-    "@ledgerhq/live-device-core": "workspace:^",
     "@ledgerhq/logs": "workspace:^",
     "@ledgerhq/native-ui": "workspace:^",
     "@ledgerhq/react-native-hid": "workspace:^",
