+++ apps/ledger-live-mobile/src/screens/Settings/Debug/Connectivity/CommandSender.tsx
@@ -7,8 +7,8 @@ import { withDevice } from "@ledgerhq/live-common/hw/deviceAccess";
 import getDeviceInfo from "@ledgerhq/live-common/hw/getDeviceInfo";
 import quitApp from "@ledgerhq/live-common/hw/quitApp";
 import getAppAndVersion from "@ledgerhq/live-common/hw/getAppAndVersion";
-import { getDeviceName } from "@ledgerhq/live-common/device/use-cases/getDeviceNameUseCase";
 import getBatteryStatus from "@ledgerhq/live-common/hw/getBatteryStatus";
+import { getDeviceName } from "@ledgerhq/live-common/device/index";
 import listApps from "@ledgerhq/live-common/hw/listApps";
 import Transport from "@ledgerhq/hw-transport";
 import { ScreenName } from "~/const";
