--- libs/ledger-live-common/src/device-core/commands/entities/FirmwareInfoEntity.ts
@@ -1,14 +1,0 @@
-export type FirmwareInfoEntity = {
-  isBootloader: boolean;
-  rawVersion: string; // if SE seVersion, if BL blVersion
-  targetId: number; // if SE seTargetId, if BL mcuTargetId
-  seVersion?: string;
-  mcuVersion: string; // NB historically not undefined. but will be ""
-  mcuBlVersion?: string;
-  mcuTargetId?: number;
-  seTargetId?: number;
-  flags: Buffer;
-  bootloaderVersion?: string;
-  hardwareVersion?: number;
-  languageId?: number;
-};
