--- libs/ledger-live-common/src/device-core/commands/use-cases/isDeviceLocalizationSupported.test.ts
@@ -1,48 +1,0 @@
-import { DeviceModelId } from "@ledgerhq/devices";
-import { isDeviceLocalizationSupported } from "./isDeviceLocalizationSupported";
-const { nanoS, nanoSP, nanoX, stax } = DeviceModelId;
-
-test("isDeviceLocalizationSupported", () => {
-  /**
-   * Nano X
-   * */
-  expect(isDeviceLocalizationSupported("2.0.0", nanoX)).toBe(false);
-  expect(isDeviceLocalizationSupported("2.0.0-whatever0", nanoX)).toBe(false);
-
-  expect(isDeviceLocalizationSupported("2.1.0", nanoX)).toBe(true);
-  expect(isDeviceLocalizationSupported("2.1.0-rc1", nanoX)).toBe(true);
-  expect(isDeviceLocalizationSupported("2.1.1", nanoX)).toBe(true);
-  expect(isDeviceLocalizationSupported("2.1.1-whatever0", nanoX)).toBe(true);
-  expect(isDeviceLocalizationSupported("2.2.0", nanoX)).toBe(true);
-  expect(isDeviceLocalizationSupported("2.2.0-whatever0", nanoX)).toBe(true);
-  expect(isDeviceLocalizationSupported("3.0.0", nanoX)).toBe(true);
-  expect(isDeviceLocalizationSupported("3.0.0-whatever0", nanoX)).toBe(true);
-
-  /**
-   * Nano SP
-   * */
-  expect(isDeviceLocalizationSupported("1.0.0", nanoSP)).toBe(false);
-  expect(isDeviceLocalizationSupported("1.0.0-whatever", nanoSP)).toBe(false);
-
-  expect(isDeviceLocalizationSupported("1.1.0", nanoSP)).toBe(true);
-  expect(isDeviceLocalizationSupported("1.1.0-rc1", nanoSP)).toBe(true);
-  expect(isDeviceLocalizationSupported("1.1.2", nanoSP)).toBe(true);
-  expect(isDeviceLocalizationSupported("1.1.2-whatever0", nanoSP)).toBe(true);
-  expect(isDeviceLocalizationSupported("1.2.0", nanoSP)).toBe(true);
-  expect(isDeviceLocalizationSupported("1.2.0-whatever0", nanoSP)).toBe(true);
-  expect(isDeviceLocalizationSupported("2.0.0", nanoSP)).toBe(true);
-  expect(isDeviceLocalizationSupported("2.0.0-whatever0", nanoSP)).toBe(true);
-
-  /**
-   * Stax
-   * */
-  expect(isDeviceLocalizationSupported("0.9.0", stax)).toBe(false);
-  expect(isDeviceLocalizationSupported("9.0.0", stax)).toBe(true);
-  expect(isDeviceLocalizationSupported("9.0.0-whatever", stax)).toBe(true);
-
-  /**
-   * NanoS
-   * */
-  expect(isDeviceLocalizationSupported("9.0.0", nanoS)).toBe(false);
-  expect(isDeviceLocalizationSupported("9.0.0-whatever", nanoS)).toBe(false);
-});
