--- libs/ledger-live-common/src/device-core/managerApi/entities/Id.ts
@@ -1,1 +1,0 @@
-export type Id = number;
