--- libs/ledger-live-common/src/device-core/managerApi/use-cases/fetchMcus.ts
@@ -1,5 +1,0 @@
-import { ManagerApiRepository } from "../repositories/ManagerApiRepository";
-
-export function fetchMcus(managerApiRepository: ManagerApiRepository) {
-  return managerApiRepository.fetchMcus();
-}
