+++ libs/ledger-live-common/src/device/index.ts
@@ -1,0 +1,13 @@
+export {
+  getDeviceName,
+  getVersion,
+  isBootloaderVersionSupported,
+  isHardwareVersionSupported,
+  isDeviceLocalizationSupported,
+  FirmwareUpdateContext,
+} from "@ledgerhq/live-device-core";
+
+export * from "./use-cases/fetchMcusUseCase";
+export * from "./use-cases/getLatestFirmwareForDeviceUseCase";
+export * from "./use-cases/listAppsUseCase";
+export * from "./hooks/useLatestFirmware";
