--- libs/ledger-live-common/src/device/use-cases/getDeviceNameUseCase.ts
@@ -1,1 +1,0 @@
-export { getDeviceName } from "@ledgerhq/live-device-core";
