--- libs/ledger-live-common/src/device/use-cases/getVersionUseCase.ts
@@ -1,5 +1,0 @@
-export {
-  getVersion,
-  isBootloaderVersionSupported,
-  isHardwareVersionSupported,
-} from "@ledgerhq/live-device-core";
