+++ libs/live-device-core/package.json
@@ -65,20 +65,5 @@
         "lib-es/*"
       ]
     }
-  },
-  "exports": {
-    "./lib/*": "./lib/*.js",
-    "./lib/*.js": "./lib/*.js",
-    "./lib-es/*": "./lib-es/*.js",
-    "./lib-es/*.js": "./lib-es/*.js",
-    "./*": {
-      "require": "./lib/*.js",
-      "default": "./lib-es/*.js"
-    },
-    "./*.js": {
-      "require": "./lib/*.js",
-      "default": "./lib-es/*.js"
-    },
-    "./package.json": "./package.json"
   }
 }
